refactor: move build_lean_project to utils
