+++ sorrydb/clients/rfl_client/rfl_client.py
@@ -8,8 +8,7 @@ from typing import Dict, Optional, Tuple
 
 from utils.git_ops import prepare_repository
 from utils.repl_ops import LeanRepl, setup_repl
-
-from database.process_sorries import build_lean_project
+from utils.lean_repo import build_lean_project
 
 # Create a module-level logger
 logger = logging.getLogger(__name__)
